+++ kogito-codegen-modules/kogito-codegen-api/src/test/java/org/kie/kogito/codegen/api/utils/KogitoContextTestUtils.java
@@ -20,6 +20,8 @@ import org.kie.kogito.codegen.api.context.impl.JavaKogitoBuildContext;
 import org.kie.kogito.codegen.api.context.impl.QuarkusKogitoBuildContext;
 import org.kie.kogito.codegen.api.context.impl.SpringBootKogitoBuildContext;
 
+import java.util.Collection;
+import java.util.function.Predicate;
 import java.util.stream.Stream;
 
 public class KogitoContextTestUtils {
@@ -34,6 +36,27 @@ public class KogitoContextTestUtils {
                 Arguments.of(QuarkusKogitoBuildContext.builder()),
                 Arguments.of(SpringBootKogitoBuildContext.builder())
         );
+    }
+
+    public static Predicate<String> mockClassAvailabilityResolver(Collection<String> includedClasses, Collection<String> excludedClasses) {
+        return mockClassAvailabilityResolver(includedClasses, excludedClasses, KogitoContextTestUtils.class.getClassLoader());
+    }
+
+    public static Predicate<String> mockClassAvailabilityResolver(Collection<String> includedClasses, Collection<String> excludedClasses, ClassLoader classLoader) {
+        return className -> {
+            if(includedClasses.contains(className)) {
+                return true;
+            }
+            else if(excludedClasses.contains(className)) {
+                return false;
+            }
+            try {
+                classLoader.loadClass(className);
+                return true;
+            } catch (ClassNotFoundException ex) {
+                return false;
+            }
+        };
     }
 
 }
