+++ kogito-codegen-modules/kogito-codegen-rules/pom.xml
@@ -34,6 +34,12 @@
 
         <!-- test -->
         <dependency>
+            <groupId>org.kie.kogito</groupId>
+            <artifactId>kogito-codegen-api</artifactId>
+            <type>test-jar</type>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
             <groupId>org.junit.jupiter</groupId>
             <artifactId>junit-jupiter-engine</artifactId>
             <scope>test</scope>
